added speed limit of 10cm per second

+++ include/RandomWalkStrategy.hpp
@@ -9,7 +9,7 @@ class RandomWalkStrategy : public IStrategy
 {
 private:
     uint64_t _turnTimeThreshold;       //seconds
-    uint64_t _randomTurnTimeThreshold; //seconds
+    float _randomTurnTimeThreshold; //seconds
     std::chrono::time_point<std::chrono::system_clock> _turnStartTime;
     std::chrono::time_point<std::chrono::system_clock> _randomTurnStartTime;
